+++ src/App.xaml.cs
@@ -1,4 +1,5 @@
-﻿using System.IO;
+﻿
+using System.IO;
 using System.Windows;
 using System.Windows.Threading;